chore: reorganize wal code a bit
